Store extension version in browser storage (#220)
This fix the issue where the update notification was shown twice to the user when browsing on github.com and gist.github.com. That happened because we stored the version in the LocalStorage. The LS is bounded to the domain and therefore the notification was appearing twice to the user. By using the browser build in storage this can be avoided.