Move path/githubClass patterns into plugin files
This helps make the plugins a little more self-contained, rather than
referring to a separate `pattern-preset.js` file. This was mostly
straight-forward, but there were two plugins that needed additional
care:

1. The `homebrew-manifest` plugin reuses the `ruby` plugin's patterns.
   To make this work, I made the `getPattern()` function static so that the
   `homebrew-manifest` plugin can reference the `ruby` plugin's patterns
   without having to create an instance of the plugin.

2. The `javascript` plugin used the `patternPresets` function to merge
   the JavaScript and CoffeeScript patterns. Instead, it now just
   directly returns the merged object. I removed the `patternPresets`
   function and tests, because it is no longer neeeded.

I also converted older `getPattern()`
implementations that just returned an array to instead return a
`{pathPatterns, githubClasses}` object. This makes the plugins more
consistent, and allows the plugin manager and tests to be simpler.