Fix openTab handler for Firefox
Before this commit, right-clicking (or middle-mouse clicking) on a link
doesn't open it in a new tab in Firefox. This was because the extension
is using a chrome-specific module for the storage.

Instead, I've changed it to use the webextension-polyfill, which makes
the storage APIs work the same in Chrome and Firefox, so you don't have
to care about platform-specific APIs.

License: MIT
Signed-off-by: Victor Bjelkholm <git@victor.earth>